Added auth-login - github
