+++ .env.example
@@ -1,1 +1,9 @@
-TOGETHER_API_KEY=
+TOGETHER_API_KEY=your_together_api_key
+GITHUB_ID=your_github_client_id
+GITHUB_SECRET=your_github_client_secret
+
+GOOGLE_CLIENT_ID=your_google_client_id
+GOOGLE_CLIENT_SECRET=your_google_client_secret
+
+NEXTAUTH_SECRET=some_random_secret
+NEXTAUTH_URL=http://localhost:3000
