+++ components/AuthProvider.tsx
@@ -1,0 +1,9 @@
+// components/AuthProvider.tsx
+"use client";
+
+import { SessionProvider } from "next-auth/react";
+import { ReactNode } from "react";
+
+export default function AuthProvider({ children }: { children: ReactNode }) {
+  return <SessionProvider>{children}</SessionProvider>;
+}
