+++ components/LogoutButton.tsx
@@ -1,0 +1,12 @@
+import { signOut } from "next-auth/react";
+
+export function LogoutButton() {
+  return (
+    <button
+      onClick={() => signOut({ callbackUrl: "/login", redirect: true })}
+      className="rounded-full border border-gray-300 bg-white px-4 py-2 text-sm text-gray-600 shadow-md transition-colors hover:bg-gray-100 dark:bg-zinc-800 dark:text-white dark:border-gray-600 dark:hover:bg-zinc-700"
+    >
+      Logout
+    </button>
+  );
+}
