+++ package.json
@@ -11,6 +11,7 @@
     "@heroicons/react": "^2.1.5",
     "@tailwindcss/forms": "^0.5.9",
     "next": "^14.2.16",
+    "next-auth": "^4.24.11",
     "next-plausible": "^3.12.2",
     "next-themes": "^0.4.6",
     "react": "18.3.1",
