+++ tsconfig.json
@@ -1,5 +1,10 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["./*"]
+    },
+    "types": ["next-auth"],
     "target": "es5",
     "lib": [
       "dom",
